Buggy Movement
Currently I was able to get my struct working such that the list holds almost a tuple of x and y values. I try to update each value on movement by the location and direction of the value in elements[i + 1] where i is an element in the array one before.... Doesnt work. but its late. So i need to ask calvin