Buggy Movement
I have done majority of this project and I understand that the understanding of c is somewhat a barrier, but effort is as well. The final thing before the project is done is to fix the buggy movement that I am 100% confident that all of you can figure out. As of right now the list works like this. It is a list of structs where each struct holds an x and y coordinate. The head of the snake is at array[0] and you can access its x and y by saying array[0].x.

The issue is as of right now is that if depending on the direction the head of the snake is moving we adjust the rest of the tail in that same direction subtracted by 1 x or y value of the element in front of it in the array. But rather than doing this, we need to think of each location in the list as a memory of the last location the element in the list i + 1 infront of it existed such that it mimics the movements of a snake. 

You then need to be able to shift all the elements down the list.